afterCheck
