+++ formInit.js
@@ -1,0 +1,5 @@
+const container = document.createElement("form");
+container.className = "form";
+document.body.append(container);
+
+export default container;
