+++ inputAge.js
@@ -1,0 +1,7 @@
+const inputAge = document.createElement("input");
+inputAge.className = "input-age";
+inputAge.type = "number";
+inputAge.size = "15";
+inputAge.placeholder = "enter your age";
+
+export default inputAge
