+++ inputName.js
@@ -1,0 +1,7 @@
+const inputName = document.createElement("input");
+inputName.className = "input-name";
+inputName.type = "text";
+inputName.placeholder = "enter your name";
+inputName.size = "20";
+
+export default inputName
